Fixes and final implementation of #6
Fixed:
- Formatting on Application layer #25
- Casting issues
- An switch logic issue that would print incorrect values for certain values in the applicaiton layer

+++ ScET/Tracer/Tracer.cpp
@@ -106,7 +106,7 @@ void Tracer::processInput(const uint8_t *buffer, const int bufferSize) {
 	case SIMTraceCommand::Data: {
 		if (bufferSize > header.totalBufferSize) { // multiple traces have been mangled together, we must separate them
 			const uint8_t *demangledBuffer = buffer + header.totalBufferSize;
-			size_t demangledBufferSize = bufferSize - header.totalBufferSize;
+			int demangledBufferSize = bufferSize - (int)header.totalBufferSize;
 			processInput(buffer, bufferSize - demangledBufferSize); // process the original trace and then the demangled one
 			return processInput(demangledBuffer, demangledBufferSize); // recursively call this in case there are more than two traces mangled together
 		} else if (bufferSize < header.totalBufferSize) { // something went wrong abort the trace
